Refactor playable
- transfer .media to playable class
- playble is its own file to avoid circular import of `media` in to `base`
